Update README.md
I am having issues figuring out how to get this program to make the username that is in both the current list and the new list unavailable if the new name is capitalized, lowered, or in all caps. I cannot figure out how exactly to write it in order for python to recognize that no matter what, that name is unavailable.